unit test is ready

+++ src/main/java/com/test/casino/repository/BetRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 import java.util.Optional;
 
 public interface BetRepository extends JpaRepository<Bet, Integer> {
-    Optional<List<Bet>> findAllByUserId(Integer userId);
+    Optional<List<Bet>> findAllByUserId(Long userId);
 }
